refactor(modules): extract prototypes/localStorage.mjs from prototype.mjs
